Add support for pre-compiled formatters and parsers
globalize-compiler allows you to pre-compile modules based on the formatters and parsers your app uses, which elminates the need to load the CLDR data at runtime, and leads to lower memory usage and faster performance.

express-globalize has now been updated to support using these compiled files in a transparent way -- just pass in the path to where the files are located, and express-globalize will load and use them instead of the full CLDR.

As part of this update, dependency versions have been updated and adjusted as necessary.